refactor, remove obsolete code
